+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - UNRELEASED
+
+### ADDED
+
+- `white-space: nowrap` added to `<code>` tags on posts
+- Proper title to Blog page
+- Proper title to Posts pages
+- Like Button on post single pages
+
+### CHANGED
+
+- "Thoughts" title changed to "Blog" on index page
+- Font size for links in the footer on mobile
+
 ## [1.3.2] - 12/01/2019
 
 ### ADDED